+++ src/msgbuf/msgbuf_head.h
@@ -113,7 +113,7 @@ static inline void msgbuf_head_decr_waiters (struct msgbuf_head *bh)
 
 int msgbuf_head_preinstall_iovs (struct msgbuf_head *bh, struct rex_iov *iovs, int n);
 
-int msgbuf_head_install_iovs (struct msgbuf_head *bh, struct rex_iov *iovs, u32 length);
+int msgbuf_head_install_iovs (struct msgbuf_head *bh, struct rex_iov *iovs, i64 length);
